Agents_ADK/requirements.txt
google-adk>=0.2.1
google-cloud-aiplatform>=1.59.0
google-cloud-firestore>=2.16.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.2
pydantic>=2.7.0
numpy>=1.26.0
scikit-learn>=1.4.2
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0

